+++ standalone/preload.js
@@ -1,6 +1,7 @@
 const { contextBridge, ipcRenderer, webUtils } = require('electron');
 
 const editorWorkerUrl = new URL('../media/editor.worker.js', window.location.href).toString();
+const diffWorkerUrl = new URL('../media/diff.worker.js', window.location.href).toString();
 
 ipcRenderer.on('bygone:host-message', (_event, payload) => {
     window.postMessage({
@@ -11,6 +12,7 @@ ipcRenderer.on('bygone:host-message', (_event, payload) => {
 contextBridge.exposeInMainWorld('__BYGONE_HOST__', {
     environment: 'standalone',
     editorWorkerUrl,
+    diffWorkerUrl,
     getPathForFile(file) {
         return webUtils.getPathForFile(file);
     },